add collection

+++ src/GiftBox/Lollipop.java
@@ -3,7 +3,7 @@ package GiftBox;
 public class Lollipop extends Sweet {
 
 
-    public Lollipop(String name, Integer count, double price, int weight) {
+    public Lollipop(String name, Integer count, double price, Double weight) {
         super(name, count, price, weight);
     }
 }
